Added music
There is a framework in there for adding a volume scale, but I couldn't get it to work. The main parts are GameManager.cs in Awake(), Update(), Game Settings, and line 123. Same thing with lines 22 in iggy.cs and the related lines in barel.cd, kong.cs, player.cs, and musicplayer.cs